refactor(eval): extract pipeline.rs
